requirements.txt
django==5.2.1
djangorestframework
opencv-python
ultralytics
numpy
torch
